+++ test/basic_crawler.js
@@ -11,7 +11,7 @@ import { LOCAL_EMULATION_DIR } from './_helper';
 chai.use(chaiAsPromised);
 
 describe('BasicCrawler', () => {
-    it('runs in parallel thru all the requests', async () => {
+    it('should run in parallel thru all the requests', async () => {
         const startedAt = Date.now();
         const sources = _.range(0, 500).map(index => ({ url: `https://example.com/${index}` }));
 
@@ -38,7 +38,7 @@ describe('BasicCrawler', () => {
         expect(await requestList.isEmpty()).to.be.eql(true);
     });
 
-    it('retries failed requests', async () => {
+    it('should retry failed requests', async () => {
         const sources = [
             { url: 'http://example.com/1' },
             { url: 'http://example.com/2' },
@@ -79,6 +79,56 @@ describe('BasicCrawler', () => {
         expect(processed['http://example.com/2'].userData.foo).to.be.a('undefined');
         expect(processed['http://example.com/2'].errorMessages).to.have.lengthOf(11);
         expect(processed['http://example.com/2'].retryCount).to.be.eql(10);
+
+        expect(await requestList.isFinished()).to.be.eql(true);
+        expect(await requestList.isEmpty()).to.be.eql(true);
+    });
+
+    it('should not retry requests with ignoreErrors set to true ', async () => {
+        const sources = [
+            { url: 'http://example.com/1', ignoreErrors: true },
+            { url: 'http://example.com/2' },
+            { url: 'http://example.com/3', ignoreErrors: true },
+        ];
+        const processed = {};
+        const requestList = new Apify.RequestList({ sources });
+
+        const handleRequestFunction = async ({ request }) => {
+            await delayPromise(10);
+            processed[request.url] = request;
+            request.userData.foo = 'bar';
+            throw Error(`This is ${request.retryCount}th error!`);
+        };
+
+        let handleFailedRequestFunctionCalls = 0;
+        const handleFailedRequestFunction = () => {
+            handleFailedRequestFunctionCalls++;
+        };
+
+        const basicCrawler = new Apify.BasicCrawler({
+            requestList,
+            maxRequestRetries: 10,
+            minConcurrency: 3,
+            maxConcurrency: 3,
+            handleRequestFunction,
+            handleFailedRequestFunction,
+        });
+
+        await requestList.initialize();
+        await basicCrawler.run();
+
+        expect(processed['http://example.com/1'].userData.foo).to.be.eql('bar');
+        expect(processed['http://example.com/1'].errorMessages).to.be.a('null');
+        expect(processed['http://example.com/1'].retryCount).to.be.eql(0);
+        expect(processed['http://example.com/3'].userData.foo).to.be.eql('bar');
+        expect(processed['http://example.com/3'].errorMessages).to.be.a('null');
+        expect(processed['http://example.com/3'].retryCount).to.be.eql(0);
+
+        expect(processed['http://example.com/2'].userData.foo).to.be.eql('bar');
+        expect(processed['http://example.com/2'].errorMessages).to.have.lengthOf(11);
+        expect(processed['http://example.com/2'].retryCount).to.be.eql(10);
+
+        expect(handleFailedRequestFunctionCalls).to.be.eql(1);
 
         expect(await requestList.isFinished()).to.be.eql(true);
         expect(await requestList.isEmpty()).to.be.eql(true);